client/fuse: add support for fuse idmapped mounts
Experimental implementation. Contains one ugly hack that
checks MDS features in a non reliable way. But it should be
easily fixable with some guidance from ceph project folks.

We already have support for idmapped mounts in the cephfs
kernel client. This is a reference implementation of
the same feature but for fuse client. It requires some
support from the libfuse side and, of course, kernel side too.

It's worth mentioning that in this commit we are enabling
idmapped mounts support by default if:
- MDS supports CEPHFS_FEATURE_HAS_OWNER_UIDGID
- kernel supports FUSE_CAP_ALLOW_IDMAP
- fuse_default_permissions = true (means that all permission
checks are done on the kernel side)

fuse_default_permissions can be set in /etc/ceph/ceph.conf:
[client]
fuse_default_permissions = true

Some extra links and examples:

- kernel support
https://github.com/mihalicyn/linux/commits/fuse_idmapped_mounts

- libfuse support
https://github.com/mihalicyn/libfuse/commits/idmap_support

Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>